+++ tests/scratch.py
@@ -1,5 +1,0 @@
-"""Scratch."""
-
-from autostorage import Database
-
-db = Database("ident.db")